+++ odoo/local-src/mtsmte_reports/models/__init__.py
@@ -2,3 +2,4 @@
 from . import company
 from . import project_task
 from . import helpers
+from . import account_invoice